requirements.txt
# Study Pilot AI - Backend Dependencies
# Offline-first, CPU-only compatible

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
# sqlite3 is built-in

# RAG Pipeline
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Document Parsing
PyPDF2==3.0.1
python-pptx==0.6.23

# Utilities
numpy==1.24.3
werkzeug==3.0.1
